Introduced a cmake file. This is a "proper" build system rather than relying on a code::blocks project file. CMake is capable of generating relevent project files for may different IDE's (including visual C++) as well as generic makefiles. As a bonus, cmakefiles are very easy to read (being mostly just a list of libraries, targets, and source files) and cmake has explicit suport for wxWidgets.
Also, partial sucess with compiling on linux has been achived. Everything but wxProgrid and icons seem to work.
I had to increase the interval on the openGL timer, as it appears that when you set a timmer to return imiedatly in
linux, linux assumes you mean it and you tend to thrash the CPU a lot.  




git-svn-id: https://gorion.oerc.ox.ac.uk:8443/svn/spinach_gui@50 fbaf8dc8-a59a-9f45-831a-ff29167b0329