This commit goes some of the way towards fixing the visualisation of tensors. See below.
The procedure for visualisating the interaction tensors is to scale them in the eigenframe and
then rotate them into the lab frame, i.e.

x^\prime = RSx

Where x is a vector on the unit sphere, R is a rotation matrix and S is the diagonal scaleing matrix.
What was actually being done was this:

x^\prime = RSR^{-1}x

This sucessfully transforms the tensor components from the eigenframe to the lab frame, which
is what you usually would want to do, but now we are assuming the unit sphere is sitting in the
lab frame when actually it's in the eigenframe. What we want is:

x^\prime = RSx