Very significant step towards a consistent unit system which is enforced by the compiler. The spinsys.hpp API has been modified to use dimension-aware types and all other files have been made consistent with these changes. There is still work to do, namely, matricese do not currently come with units.
Nearly every source file is affected. XML is current disabled due to its complexity. CASTEP quantites in ppm are not properly treated yet.
Application compiles and runs, GUI elements should be displaying things in units of MHz, but the GUI has not been updated to reflect this and
the numbers are yet to be independently verified to be correct.